requirements.txt
roboflow 
supervision 
ultralytics
python-dotenv
opencv-python
numpy
imutils
torch
pillow